pyproject.toml
[tool.poetry]
name = "flask-scaffold-clean-arch"
version = "0.1.0"
description = "A clean architecture Flask API with SOLID principles and Dockerized test/prod environments"
authors = ["Denis Candido <denisb411@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3"
flask-sqlalchemy = "^3.1"
flasgger = "^0.9.7.1"
pydantic = {extras = ["email"], version = "^2.11.7"}
dependency-injector = "^4.41"
psycopg2-binary = "^2.9.10"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4"
isort = "^5.13"
mypy = "^1.10"

[tool.poetry.group.test.dependencies]
pytest = "^8.2"
pytest-cov = "^5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
